python/agents/machine-learning-engineering/pyproject.toml
[project]
name = "machine-learning-engineering"
version = "0.1.0"
description = "MLE-STAR: A novel machine learning engineering (MLE) agent to automate the implementation of ML models."
authors = [
    { name = "Jiefeng Chen", email = "jiefengc@google.com" },
    { name = "Raj Sinha", email = "sinharaj@google.com" },
    { name = "Jaehyun Nam", email = "jaehyunnam@google.com" },
    { name = "Jinsung Yoon", email = "jinsungyoon@google.com" }
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "google-adk>=1.5.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "numpy>=2.2.3",
    "pandas>=2.3.1",
    "scikit-learn>=1.7.1",
    # Note: If using uv or pip, use --extra-index-url https://download.pytorch.org/whl/cpu 
    # to ensure the CPU version of torch is prioritized.
    "torch>=2.7.1",
    "kaggle (<1.7)",
    "mlebench @ file:///usr/local/google/home/luiskaiser/Documents/morph-adk-samples/python/agents/machine-learning-engineering/testing/temp_bench",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "google-adk[eval]>=1.5.0",
    "google-cloud-aiplatform[evaluation]>=1.93.0",
]
deployment = [
    "absl-py>=2.2.1",
]

[tool.ruff]
extend = "../../../pyproject.toml"

[tool.ruff.lint]
ignore = ["C901", "PLR0911", "PLR0912", "PLR0915"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["machine_learning_engineering"]

[tool.pytest.ini_options]
asyncio_mode = "auto"


[tool.poetry.dependencies]
mlebench = {develop = true}
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"